fix(backend): satisfy architecture boundaries
