+++ src/backend/src/agentclaw/community/core/expert_chat/README.md
@@ -13,12 +13,14 @@ consumes:
   - "BotRepository"
   - "DeviceService + DeviceContextResolver"
   - "ServiceBot repo + BaasService"
+  - "CommonConfigService (legacy ARKA image policy resolution)"
   - "SkillSync guard"
   - "CollaboratorService (chat 权限校验)"
 internal_dependencies:
   - agentclaw.community.core.bot_collaborator
   - agentclaw.community.core.bot_management
   - agentclaw.community.core.caller_identity
+  - agentclaw.community.core.common_config
   - agentclaw.community.core.devices
   - agentclaw.community.core.service_bot
   - agentclaw.community.core.skill_center
